refector: [feat/#68/refector] 리팩토링
