+++ a_uction/src/main/java/com/example/a_uction/controller/user/UserLoginController.java
@@ -1,8 +1,8 @@
-package com.example.a_uction.controller;
+package com.example.a_uction.controller.user;
 
 import com.example.a_uction.model.user.dto.LoginUser;
 import com.example.a_uction.security.jwt.dto.TokenDto;
-import com.example.a_uction.service.UserLoginService;
+import com.example.a_uction.service.user.UserLoginService;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
