+++ a_uction/src/test/java/com/example/a_uction/controller/user/UserRegisterControllerTest.java
@@ -1,4 +1,4 @@
-package com.example.a_uction.controller;
+package com.example.a_uction.controller.user;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
@@ -14,7 +14,8 @@ import com.example.a_uction.exception.AuctionException;
 import com.example.a_uction.exception.constants.ErrorCode;
 import com.example.a_uction.model.user.dto.RegisterUser;
 import com.example.a_uction.model.user.dto.Verify;
-import com.example.a_uction.service.UserRegisterService;
+import com.example.a_uction.security.jwt.JwtProvider;
+import com.example.a_uction.service.user.UserRegisterService;
 import com.example.a_uction.service.user.VerifyService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.DisplayName;
@@ -38,6 +39,8 @@ class UserRegisterControllerTest {
 	private UserRegisterService userRegisterService;
 	@MockBean
 	private VerifyService verifyService;
+	@MockBean
+	private JwtProvider jwtProvider;
 
 	@Autowired
 	private ObjectMapper objectMapper;
