+++ auction-api.http
@@ -1,7 +1,7 @@
 ###경매 생성
-POST http://localhost:8081/auction
+POST http://localhost:8081/auctions
 Content-Type: application/json
-Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJ1c2VyRW1haWwiOiJ6ZXJvYmFzZUBnbWFpbC5jb20iLCJzdWIiOiJBY2Nlc3NUb2tlbiIsImlhdCI6MTY4MTEzMTEwNywiZXhwIjoxNjgxMTM0NzA3fQ.f2WncSBZuGdcwgdLMtsVePKElVxbMfvZrcKq6194LKM
+Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJ1c2VyRW1haWwiOiJ6ZXJvYmFzZUBnbWFpbC5jb20iLCJzdWIiOiJBY2Nlc3NUb2tlbiIsImlhdCI6MTY4MTI4OTE5NywiZXhwIjoxNjgxMjkyNzk3fQ.nh01Laz3Zrj2RiZdarN4ETRb_4rZgp6MY9MPegDRF3E
 
 {
   "itemName": "air force 1",
@@ -9,17 +9,17 @@ Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJ1c2VyRW1haWwiOiJ6ZXJvYmFzZUBnbWFpb
   "startingPrice": 1000,
   "minimumBid": 100,
   "category": "SHOES",
-  "startDateTime": "2023-04-11T13:41:30.416Z",
-  "endDateTime": "2023-04-30T13:41:30.416Z"
+  "startDateTime": "2024-04-11T13:41:30.416Z",
+  "endDateTime": "2024-04-30T13:41:30.416Z"
 }
 
 ###경매 보기
-GET http://localhost:8081/auction/1
+GET http://localhost:8081/auctions/1
 Content-Type: application/json
 Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJ1c2VyRW1haWwiOiJ6ZXJvYmFzZUBnbWFpbC5jb20iLCJzdWIiOiJBY2Nlc3NUb2tlbiIsImlhdCI6MTY4MTEzMTAwNiwiZXhwIjoxNjgxMTM0NjA2fQ.1qIxzvIRdxMITQ1TheoADMls9irD44MQudfeaxETkhg
 
 ###경매 수정
-PUT http://localhost:8081/auction?auctionId=1
+PUT http://localhost:8081/auctions?auctionId=1
 Content-Type: application/json
 Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJ1c2VyRW1haWwiOiJ6ZXJvYmFzZUBnbWFpbC5jb20iLCJzdWIiOiJBY2Nlc3NUb2tlbiIsImlhdCI6MTY4MTEzMTEwNywiZXhwIjoxNjgxMTM0NzA3fQ.f2WncSBZuGdcwgdLMtsVePKElVxbMfvZrcKq6194LKM
 
@@ -34,11 +34,11 @@ Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJ1c2VyRW1haWwiOiJ6ZXJvYmFzZUBnbWFpb
 }
 
 ###user의 모든 경매 가져오기
-GET http://localhost:8081/auction/read
+GET http://localhost:8081/auctions/read
 Content-Type: application/json
 Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJ1c2VyRW1haWwiOiJ6ZXJvYmFzZUBnbWFpbC5jb20iLCJzdWIiOiJBY2Nlc3NUb2tlbiIsImlhdCI6MTY4MTEzMTEwNywiZXhwIjoxNjgxMTM0NzA3fQ.f2WncSBZuGdcwgdLMtsVePKElVxbMfvZrcKq6194LKM
 
 ###경매 삭제
-DELETE http://localhost:8081/auction/1
+DELETE http://localhost:8081/auctions/1
 Content-Type: application/json
 Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJ1c2VyRW1haWwiOiJ6ZXJvYmFzZUBnbWFpbC5jb20iLCJzdWIiOiJBY2Nlc3NUb2tlbiIsImlhdCI6MTY4MTEzMTAwNiwiZXhwIjoxNjgxMTM0NjA2fQ.1qIxzvIRdxMITQ1TheoADMls9irD44MQudfeaxETkhg
